feat: add #initialize & #valid_move? methods
Starting to code the Connect Four project using TDD
means beginning with the Board class because there
isn't much to test in the Player class (which will
likely only have #initialize).

Starting with a simple method in the Board class
after #initialize by defining a #valid_move?
method to pass the tests in board_spec.rb.